Moving out context outside of BibPage
